requirements.txt
Flask==3.0.0
pymongo==4.6.1
Flask-Session==0.5.0
python-dotenv==1.0.0
gunicorn==21.2.0
bcrypt==4.1.2
email-validator==2.1.0
